+++ plugin/src/main/kotlin/org/jlleitschuh/gradle/ktlint/KtlintPlugin.kt
@@ -238,7 +238,6 @@ open class KtlintPlugin : Plugin<Project> {
         kotlinSourceDirectories: Iterable<*>
     ): Task {
         return target.taskHelper<KtlintFormatTask>("ktlint${sourceSetName.capitalize()}Format") {
-            group = FORMATTING_GROUP
             description = "Runs a check against all .kt files to ensure that they are formatted according to ktlint."
             configurePluginTask(extension, ktLintConfig, kotlinSourceDirectories)
         }
@@ -252,7 +251,6 @@ open class KtlintPlugin : Plugin<Project> {
         kotlinSourceDirectories: Iterable<*>
     ): Task {
         return target.taskHelper<KtlintCheckTask>("ktlint${sourceSetName.capitalize()}Check") {
-            group = VERIFICATION_GROUP
             description = "Runs a check against all .kt files to ensure that they are formatted according to ktlint."
             configurePluginTask(extension, ktLintConfig, kotlinSourceDirectories)
         }