+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [6.1.0] - Unreleased
 ### Added
-  - ?
+  - Console colored output (#50)
 ### Changed
   - Update Kotlin to `1.2.71` version
   - Update Gradle to `4.10.2` version