+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
   - Update Kotlin to `1.2.71` version
   - Update Gradle to `4.10.2` version
-  - Update default KtLint version to `0.27.0`
+  - Update default KtLint version to `0.29.0`
 ### Removed
   - ?
 ### Fixed